--- types/uikit/components/countdown.d.ts
@@ -1,12 +1,0 @@
-import { UIkitElement } from "../utils";
-
-type UIkitCountdownOptions = {
-    date?: string | boolean;
-}
-
-interface UIkitCountdownElement {
-    start(): void;
-    stop(): void;
-}
-
-export type Countdown = (element: UIkitElement, options?: UIkitCountdownOptions) => UIkitCountdownElement;
