--- types/uikit/components/filter.d.ts
@@ -1,8 +1,0 @@
-import { UIkitElement } from "../utils";
-
-type UIkitFilterOptions = {
-    target?: string;
-    selActive?: string | boolean;
-}
-
-export type Filter = (element: UIkitElement, options?: UIkitFilterOptions) => void;
