--- types/uikit/components/tooltip.d.ts
@@ -1,18 +1,0 @@
-import { UIkitElement } from "../utils";
-
-type UIkitTooltipOptions = {
-    title?: string;	
-    pos?: string;
-    offset?: number | boolean;
-    animation?: string;	
-    duration?: number;	
-    delay?: number;	
-    cls?: string;
-}
-
-type UIkitTooltipElement = {
-    show(): void;
-    hide(): void;
-}
-
-export type Tooltip = (element: UIkitElement, options?: UIkitTooltipOptions) => UIkitTooltipElement;
