--- types/uikit/core/accordion.d.ts
@@ -1,19 +1,0 @@
-import { UIkitElement, UIkitNode } from "../utils";
-
-export type UIkitAccordionOptions = {
-    active?: number;
-    animation?: boolean;
-    collapsible?: boolean;	
-    content?: string;	
-    duration?: number;
-    multiple?: boolean;
-    targets?: string;
-    toggle?: string;
-    transition?: string;
-}
-
-interface UIkitAccordionElement {
-    toggle(index: string | number | UIkitNode, animate: boolean): void;
-}
-
-export type Accordion = (element: UIkitElement, options?: UIkitAccordionOptions) => UIkitAccordionElement;
