--- types/uikit/core/cover.d.ts
@@ -1,9 +1,0 @@
-import { UIkitElement } from "../utils";
-
-type UIkitCoverOptions = {
-    automute?: boolean;
-    width?: number;
-    height?: number;
-}
-
-type Cover = (element: UIkitElement, options?: UIkitCoverOptions) => void;
