--- types/uikit/core/form.d.ts
@@ -1,7 +1,0 @@
-import { UIkitElement } from "../utils";
-
-type UIkitFormOptions = {
-    target?: string | boolean;
-}
-
-type FormCustom = (element: UIkitElement, options?: UIkitFormOptions) => void;
