--- types/uikit/core/height-match.d.ts
@@ -1,8 +1,0 @@
-import { UIkitElement } from "../utils";
-
-type UIkitHeightMatchOptions = {
-    target?: string;
-    row?: boolean;
-}
-
-type HeightMatch = (element: UIkitElement, options?: UIkitHeightMatchOptions) => void;
