--- types/uikit/core/icon.d.ts
@@ -1,10 +1,0 @@
-import { UIkitElement } from "../utils";
-
-type UIkitIconOptions = {
-    icon?: string;
-    ratio?: number;
-}
-
-type Icon = (element: UIkitElement, options?: UIkitIconOptions) => {
-    svg: Promise<any>;
-};
