--- types/uikit/core/leader.d.ts
@@ -1,8 +1,0 @@
-import { UIkitElement } from "../utils";
-
-type UIkitLeaderOptions = {
-    fill?: string;
-    media?: number | string;
-}
-
-type Leader = (element: UIkitElement, options?: UIkitLeaderOptions) => void;
