--- types/uikit/core/margin.d.ts
@@ -1,8 +1,0 @@
-import { UIkitElement } from "../utils";
-
-type UIkitMarginOptions = {
-    margin?: string;
-    'first-column'?: string;
-}
-
-type Margin = (element: UIkitElement, options?: UIkitMarginOptions) => void;
