--- types/uikit/core/scroll.d.ts
@@ -1,12 +1,0 @@
-import { UIkitElement, UIkitNode } from "../utils";
-
-type UIkitScrollOptions = {
-    duration?: number;
-    offset?: number;
-}
-
-interface UIkitScrollElement {
-    scrollTo(el: string | UIkitNode): void;
-}
-
-type Scroll = (element: UIkitElement, options?: UIkitScrollOptions) => UIkitScrollElement;
