--- types/uikit/core/video.d.ts
@@ -1,8 +1,0 @@
-import { UIkitElement } from "../utils";
-
-type UIkitVideoOptions = {
-    autoplay?: boolean | string;
-    automute?: boolean;
-}
-
-type Video = (element: UIkitElement, options?: UIkitVideoOptions) => void;
