--- types/uikit/utils/index.d.ts
@@ -1,3 +1,0 @@
-export type UIkitElement = object | HTMLElement | Array<UIkitElement> | JQuery
-export type UIkitNodes = NodeList | HTMLCollection | Array<UIkitNode>
-export type UIkitNode = Node
